+++ src/bootstrap_cache.rs
@@ -9,14 +9,20 @@
 
 #![allow(unused)]
 
-use crate::utils;
-use crate::{Error, Result};
-use log::{info, warn};
+use crate::error::{Error, Result};
+use bincode::{deserialize_from, serialize_into};
+use dirs_next::home_dir;
+use log::warn;
+use serde::{de::DeserializeOwned, Serialize};
 use std::{
     collections::{HashSet, VecDeque},
-    fs, io,
+    fs,
     net::SocketAddr,
     path::{Path, PathBuf},
+};
+use std::{
+    fs::File,
+    io::{BufReader, BufWriter},
 };
 
 /// Maximum peers in the cache.
@@ -42,21 +48,16 @@ impl BootstrapCache {
     pub fn new(
         hard_coded_contacts: HashSet<SocketAddr>,
         user_override: Option<&PathBuf>,
-        fresh: bool,
     ) -> Result<BootstrapCache> {
         let get_cache_path = |dir: &PathBuf| dir.join("bootstrap_cache");
 
         let cache_path = user_override.map_or_else(
-            || Ok::<_, Error>(get_cache_path(&utils::project_dir()?)),
+            || Ok::<_, Error>(get_cache_path(&project_dir()?)),
             |d| Ok(get_cache_path(d)),
         )?;
 
         let peers = if cache_path.exists() {
-            if fresh {
-                VecDeque::new()
-            } else {
-                utils::read_from_disk(&cache_path)?
-            }
+            read_from_disk(&cache_path)?
         } else {
             let cache_dir = cache_path.parent().ok_or_else(|| {
                 Error::InvalidPath(format!(
@@ -115,7 +116,7 @@ impl BootstrapCache {
         let get_cache_path = |dir: &PathBuf| dir.join("bootstrap_cache");
 
         let cache_path = user_override.map_or_else(
-            || Ok::<_, Error>(get_cache_path(&utils::project_dir()?)),
+            || Ok::<_, Error>(get_cache_path(&project_dir()?)),
             |d| Ok(get_cache_path(d)),
         )?;
 
@@ -135,12 +136,69 @@ impl BootstrapCache {
     /// Write cached peers to disk every 10 inserted peers.
     fn try_sync_to_disk(&mut self) {
         if self.add_count > 9 {
-            if let Err(e) = utils::write_to_disk(&self.cache_path, &self.peers) {
+            if let Err(e) = write_to_disk(&self.cache_path, &self.peers) {
                 warn!("Failed to write bootstrap cache to disk: {}", e);
             }
             self.add_count = 0;
         }
     }
+}
+
+// Private utilities/helpers below
+
+/// Get the project directory
+#[cfg(any(
+    all(
+        unix,
+        not(any(target_os = "android", target_os = "androideabi", target_os = "ios"))
+    ),
+    windows
+))]
+#[inline]
+fn project_dir() -> Result<PathBuf> {
+    let dirs = home_dir().ok_or(Error::UserHomeDir)?;
+    let project_dir = dirs.join(".safe").join("qp2p");
+    Ok(project_dir)
+}
+
+/// Get the project directory
+#[cfg(not(any(
+    all(
+        unix,
+        not(any(target_os = "android", target_os = "androideabi", target_os = "ios"))
+    ),
+    windows
+)))]
+#[inline]
+fn project_dir() -> Result<Dirs> {
+    Err(Error::Configuration {
+        e: "No default project dir on non-desktop platforms. User must provide an override path."
+            .to_string(),
+    })
+}
+
+/// Try reading from the disk into the given structure.
+fn read_from_disk<D>(file_path: &Path) -> Result<D>
+where
+    D: DeserializeOwned,
+{
+    Ok(File::open(file_path)
+        .map_err(Into::into)
+        .map(BufReader::new)
+        .and_then(|mut rdr| deserialize_from(&mut rdr))?)
+}
+
+/// Try writing the given structure to the disk.
+fn write_to_disk<S>(file_path: &Path, s: &S) -> Result<()>
+where
+    S: Serialize,
+{
+    File::create(file_path)
+        .map_err(Into::into)
+        .map(BufWriter::new)
+        .and_then(|mut rdr| serialize_into(&mut rdr, s))?;
+
+    Ok(())
 }
 
 #[cfg(test)]
@@ -154,7 +212,7 @@ mod tests {
         #[test]
         fn when_10_peers_are_added_they_are_synced_to_disk() -> Result<(), Error> {
             let dirs = test_dirs();
-            let mut cache = BootstrapCache::new(Default::default(), Some(&dirs), false)?;
+            let mut cache = BootstrapCache::new(Default::default(), Some(&dirs))?;
 
             for _ in 0..10 {
                 cache.add_peer(rand_node_addr());
@@ -162,7 +220,7 @@ mod tests {
 
             assert_eq!(cache.peers.len(), 10);
 
-            let cache = BootstrapCache::new(Default::default(), Some(&dirs), false)?;
+            let cache = BootstrapCache::new(Default::default(), Some(&dirs))?;
             assert_eq!(cache.peers.len(), 10);
             Ok(())
         }
@@ -175,7 +233,7 @@ mod tests {
             assert!(hard_coded.insert(peer1));
 
             let dirs = test_dirs();
-            let mut cache = BootstrapCache::new(hard_coded, Some(&dirs), false)?;
+            let mut cache = BootstrapCache::new(hard_coded, Some(&dirs))?;
 
             cache.add_peer(peer1);
             cache.add_peer(peer2);
@@ -190,7 +248,7 @@ mod tests {
             let dirs = test_dirs();
             let port_base = 5000;
 
-            let mut cache = BootstrapCache::new(Default::default(), Some(&dirs), false)?;
+            let mut cache = BootstrapCache::new(Default::default(), Some(&dirs))?;
 
             for i in 0..MAX_CACHE_SIZE {
                 cache.add_peer(make_node_addr(port_base + i as u16));
@@ -209,7 +267,7 @@ mod tests {
         #[test]
         fn it_moves_given_node_to_the_top_of_the_list() -> Result<(), Error> {
             let dirs = test_dirs();
-            let mut cache = BootstrapCache::new(Default::default(), Some(&dirs), false)?;
+            let mut cache = BootstrapCache::new(Default::default(), Some(&dirs))?;
             let peer1 = rand_node_addr();
             let peer2 = rand_node_addr();
             let peer3 = rand_node_addr();